API Reference

This page lists the public API of MathOptAI.

Info

This page is an unstructured list of the MathOptAI API. For a more structured overview, read the Manual or Tutorial parts of this documentation.

Load all of the public the API into the current scope with:

using MathOptAI

Alternatively, load only the module with:

import MathOptAI

and then prefix all calls with MathOptAI. to create MathOptAI.<NAME>.

AbstractPredictor

MathOptAI.AbstractPredictor — Type

abstract type AbstractPredictor end

An abstract type representing different types of prediction models.

Methods

All subtypes must implement:

add_predictor

build_predictor

The following methods are optional, but encouraged:

output_size

add_predictor

MathOptAI.add_predictor — Function

MathOptAI.add_predictor(
    model::JuMP.AbstractModel,
    predictor::MathOptAI.Quantile{<:AbstractGPs.PosteriorGP},
    x::Vector,
)

Add the quantiles of a trained Gaussian Process from AbstractGPs.jl to model.

Example

julia> using JuMP, MathOptAI, AbstractGPs

julia> x_data = 2π .* (0.0:0.1:1.0);

julia> y_data = sin.(x_data);

julia> fx = AbstractGPs.GP(AbstractGPs.Matern32Kernel())(x_data, 0.1);

julia> p_fx = AbstractGPs.posterior(fx, y_data);

julia> model = Model();

julia> @variable(model, 1 <= x[1:1] <= 6, start = 3);

julia> predictor = MathOptAI.Quantile(p_fx, [0.1, 0.9]);

julia> y, _ = MathOptAI.add_predictor(model, predictor, x);

julia> y
2-element Vector{VariableRef}:
 moai_quantile[1]
 moai_quantile[2]

julia> @objective(model, Max, y[2] - y[1])
moai_quantile[2] - moai_quantile[1]

MathOptAI.add_predictor(
    model::JuMP.AbstractModel,
    predictor::StatsModels.TableRegressionModel,
    x::DataFrames.DataFrame;
    kwargs...,
)

Add a trained regression model from StatsModels.jl to model, using the DataFrame x as input.

In most cases, predictor should be a GLM.jl predictor supported by MathOptAI, but trained using @formula and a DataFrame instead of the raw matrix input.

In general, x may have some columns that are constant (Float64) and some columns that are JuMP decision variables.

Keyword arguments

All keyword arguments are passed to the corresponding add_predictor of the GLM extension.

Example

julia> using DataFrames, GLM, JuMP, MathOptAI

julia> train_df = DataFrames.DataFrame(x1 = rand(10), x2 = rand(10));

julia> train_df.y = 1.0 .* train_df.x1 + 2.0 .* train_df.x2 .+ rand(10);

julia> predictor = GLM.lm(GLM.@formula(y ~ x1 + x2), train_df);

julia> model = Model();

julia> test_df = DataFrames.DataFrame(
           x1 = rand(6),
           x2 = @variable(model, [1:6]),
       );

julia> test_df.y, _ = MathOptAI.add_predictor(model, predictor, test_df);

julia> test_df.y
6-element Vector{VariableRef}:
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]

add_predictor(
    model::JuMP.AbstractModel,
    predictor::Any,
    x::Vector;
    reduced_space::Bool = false,
    kwargs...,
)::Tuple{<:Vector,<:AbstractFormulation}

Return a Vector representing y such that y = predictor(x) and an AbstractFormulation containing the variables and constraints that were added to the model.

The element type of x is deliberately unspecified. The vector x may contain any mix of scalar constants, JuMP decision variables, and scalar JuMP functions like AffExpr, QuadExpr, or NonlinearExpr.

Keyword arguments

reduced_space: if true, wrap predictor in ReducedSpace before adding to the model.

All other keyword arguments are passed to build_predictor.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> f = MathOptAI.Affine([2.0, 3.0])
Affine(A, b) [input: 2, output: 1]

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
1-element Vector{VariableRef}:
 moai_Affine[1]

julia> formulation
Affine(A, b) [input: 2, output: 1]
├ variables [1]
│ └ moai_Affine[1]
└ constraints [1]
  └ 2 x[1] + 3 x[2] - moai_Affine[1] = 0

julia> y, formulation = MathOptAI.add_predictor(model, f, x; reduced_space = true);

julia> y
1-element Vector{AffExpr}:
 2 x[1] + 3 x[2]

julia> formulation
ReducedSpace(Affine(A, b) [input: 2, output: 1])
├ variables [0]
└ constraints [0]

add_predictor(model::JuMP.AbstractModel, predictor, x::Array; kwargs...)

This method is a helper function for adding predictor to model when the input x is a multi-dimensional array.

It is equivalent to passing vec(x) with the keyword input_size = size(x).

Example

julia> using JuMP, MathOptAI, Flux

julia> model = Model();

julia> @variable(model, x[1:4, 1:4]);

julia> predictor = Flux.Chain(Flux.MaxPool((2, 2)), Flux.flatten);

julia> y, formulation = MathOptAI.add_predictor(model, predictor, x);

julia> y
4-element Vector{VariableRef}:
 moai_MaxPool2d[1]
 moai_MaxPool2d[2]
 moai_MaxPool2d[3]
 moai_MaxPool2d[4]

julia> formulation
MaxPool2d((4, 4, 1), (2, 2), (2, 2), (0, 0))
├ variables [4]
│ ├ moai_MaxPool2d[1]
│ ├ moai_MaxPool2d[2]
│ ├ moai_MaxPool2d[3]
│ └ moai_MaxPool2d[4]
└ constraints [4]
  ├ moai_MaxPool2d[1] - max(max(max(x[1,1], x[2,1]), x[1,2]), x[2,2]) = 0
  ├ moai_MaxPool2d[2] - max(max(max(x[3,1], x[4,1]), x[3,2]), x[4,2]) = 0
  ├ moai_MaxPool2d[3] - max(max(max(x[1,3], x[2,3]), x[1,4]), x[2,4]) = 0
  └ moai_MaxPool2d[4] - max(max(max(x[3,3], x[4,3]), x[3,4]), x[4,4]) = 0

build_predictor

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(predictor::EvoTrees.EvoTree{L,1}) where {L}

Convert a boosted tree from EvoTrees.jl to an AffineCombination of BinaryDecisionTree.

Example

julia> using JuMP, MathOptAI, EvoTrees

julia> truth(x::Vector) = x[1] <= 0.5 ? -2 : (x[2] <= 0.3 ? 3 : 4)
truth (generic function with 1 method)

julia> x_train = abs.(sin.((1:10) .* (3:4)'));

julia> size(x_train)
(10, 2)

julia> y_train = truth.(Vector.(eachrow(x_train)));

julia> config = EvoTrees.EvoTreeRegressor(; nrounds = 3);

julia> tree = EvoTrees.fit(config; x_train, y_train);

julia> model = Model();

julia> @variable(model, 0 <= x[1:2] <= 1);

julia> y, _ = MathOptAI.add_predictor(model, tree, x);

julia> y
1-element Vector{VariableRef}:
 moai_AffineCombination[1]

julia> MathOptAI.build_predictor(tree)
AffineCombination
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
└ 1.0 * [2.0]

output_size

MathOptAI.output_size — Function

output_size(predictor::AbstractPredictor, input_size::Nothing)
output_size(predictor::AbstractPredictor, input_size::NTuple{N,Int}) where {N}

Return the output size of predictor with an input with shape input_size.

If input_size === nothing, no information about the input is known. This function returns an NTuple{N,Int} if a static output size based on the predictor, otherwise returns nothing.

Example

julia> using MathOptAI

julia> output_size(ReLU(), nothing)

julia> output_size(ReLU(), (2,))
(2,)

julia> output_size(MaxPool2d((3, 3); input_size = (6, 9, 1)), (6, 9, 1))
(2, 3, 1)

Affine

MathOptAI.Affine — Type

Affine(
    A::Matrix{T},
    b::Vector{T} = zeros(T, size(A, 1)),
) where {T} <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = A x + b\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, 0 <= x[i in 1:2] <= i);

julia> f = MathOptAI.Affine([2.0 3.0], [4.0])
Affine(A, b) [input: 2, output: 1]

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
1-element Vector{VariableRef}:
 moai_Affine[1]

julia> formulation
Affine(A, b) [input: 2, output: 1]
├ variables [1]
│ └ moai_Affine[1]
└ constraints [3]
  ├ moai_Affine[1] ≥ 4
  ├ moai_Affine[1] ≤ 12
  └ 2 x[1] + 3 x[2] - moai_Affine[1] = -4

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
1-element Vector{AffExpr}:
 2 x[1] + 3 x[2] + 4

julia> formulation
ReducedSpace(Affine(A, b) [input: 2, output: 1])
├ variables [0]
└ constraints [0]

AffineCombination

MathOptAI.AffineCombination — Type

AffineCombination(
    predictors::Vector{<:AbstractPredictor},
    weights::Vector{Float64},
    constant::Vector{Float64},
)

An AbstractPredictor that represents the linear combination of other predictors.

The main purpose of this predictor is to model random forests and gradient boosted trees.

A random forest is the mean a set of BinaryDecisionTree

A gradient boosted tree is the sum of a set of BinaryDecisionTree

Example

julia> using JuMP, MathOptAI

julia> rhs = MathOptAI.BinaryDecisionTree(1, 1.0, 0, 1)
BinaryDecisionTree{Float64,Int64} [leaves=2, depth=2]

julia> lhs = MathOptAI.BinaryDecisionTree(1, -0.1, -1, 0)
BinaryDecisionTree{Float64,Int64} [leaves=2, depth=2]

julia> tree_1 = MathOptAI.BinaryDecisionTree(1, 0.0, -1, rhs);

julia> tree_2 = MathOptAI.BinaryDecisionTree(1, 0.9, lhs, 1);

julia> random_forest = MathOptAI.AffineCombination(
           [tree_1, tree_2],
           [0.5, 0.5],
           [0.0],
       )
AffineCombination
├ 0.5 * BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
├ 0.5 * BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
└ 1.0 * [0.0]

julia> model = Model();

julia> @variable(model, -3 <= x[1:1] <= 5);

julia> y, formulation = MathOptAI.add_predictor(model, random_forest, x);

julia> y
1-element Vector{VariableRef}:
 moai_AffineCombination[1]

julia> formulation
AffineCombination
├ 0.5 * BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
├ 0.5 * BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
└ 1.0 * [0.0]
├ variables [1]
│ └ moai_AffineCombination[1]
└ constraints [1]
  └ 0.5 moai_BinaryDecisionTree_value[1] + 0.5 moai_BinaryDecisionTree_value[1] - moai_AffineCombination[1] = 0
BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
├ variables [4]
│ ├ moai_BinaryDecisionTree_value[1]
│ ├ moai_BinaryDecisionTree_z[1]
│ ├ moai_BinaryDecisionTree_z[2]
│ └ moai_BinaryDecisionTree_z[3]
└ constraints [7]
  ├ moai_BinaryDecisionTree_z[1] + moai_BinaryDecisionTree_z[2] + moai_BinaryDecisionTree_z[3] = 1
  ├ moai_BinaryDecisionTree_z[1] --> {x[1] ≤ -1.0e-6}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≥ 0}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≤ 0.999999}
  ├ moai_BinaryDecisionTree_z[3] --> {x[1] ≥ 0}
  ├ moai_BinaryDecisionTree_z[3] --> {x[1] ≥ 1}
  └ moai_BinaryDecisionTree_z[1] - moai_BinaryDecisionTree_z[3] + moai_BinaryDecisionTree_value[1] = 0
BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
├ variables [4]
│ ├ moai_BinaryDecisionTree_value[1]
│ ├ moai_BinaryDecisionTree_z[1]
│ ├ moai_BinaryDecisionTree_z[2]
│ └ moai_BinaryDecisionTree_z[3]
└ constraints [7]
  ├ moai_BinaryDecisionTree_z[1] + moai_BinaryDecisionTree_z[2] + moai_BinaryDecisionTree_z[3] = 1
  ├ moai_BinaryDecisionTree_z[1] --> {x[1] ≤ 0.899999}
  ├ moai_BinaryDecisionTree_z[1] --> {x[1] ≤ -0.100001}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≤ 0.899999}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≥ -0.1}
  ├ moai_BinaryDecisionTree_z[3] --> {x[1] ≥ 0.9}
  └ moai_BinaryDecisionTree_z[1] - moai_BinaryDecisionTree_z[3] + moai_BinaryDecisionTree_value[1] = 0

AvgPool2d

MathOptAI.AvgPool2d — Type

AvgPool2d(
    kernel_size::Tuple{Int,Int};
    input_size::Tuple{Int,Int,Int},
    stride::Tuple{Int,Int} = (1, 1),
    padding::Tuple{Int,Int} = (0, 0),
) <: AbstractPredictor

An AbstractPredictor that represents a mean pooling layer.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[h in 1:2, w in 1:4])
2×4 Matrix{VariableRef}:
 x[1,1]  x[1,2]  x[1,3]  x[1,4]
 x[2,1]  x[2,2]  x[2,3]  x[2,4]

julia> predictor = MathOptAI.AvgPool2d((2, 2); input_size = (2, 4, 1))
AvgPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0))

julia> y, formulation = MathOptAI.add_predictor(model, predictor, vec(x));

julia> y
2-element Vector{VariableRef}:
 moai_AvgPool2d[1]
 moai_AvgPool2d[2]

julia> formulation
AvgPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0))
├ variables [2]
│ ├ moai_AvgPool2d[1]
│ └ moai_AvgPool2d[2]
└ constraints [2]
  ├ -0.25 x[1,1] - 0.25 x[2,1] - 0.25 x[1,2] - 0.25 x[2,2] + moai_AvgPool2d[1] = 0
  └ -0.25 x[1,3] - 0.25 x[2,3] - 0.25 x[1,4] - 0.25 x[2,4] + moai_AvgPool2d[2] = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, predictor, vec(x); reduced_space = true);

julia> y
2-element Vector{AffExpr}:
 0.25 x[1,1] + 0.25 x[1,2] + 0.25 x[2,1] + 0.25 x[2,2]
 0.25 x[1,3] + 0.25 x[1,4] + 0.25 x[2,3] + 0.25 x[2,4]

julia> formulation
ReducedSpace(AvgPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0)))
├ variables [0]
└ constraints [0]

BinaryDecisionTree

MathOptAI.BinaryDecisionTree — Type

BinaryDecisionTree{K,V}(
    feat_id::Int,
    feat_value::K,
    lhs::Union{V,BinaryDecisionTree{K,V}},
    rhs::Union{V,BinaryDecisionTree{K,V}},
    atol::Float64 = 1e-6,
)

An AbstractPredictor that represents a binary decision tree.

If x[feat_id] <= feat_value - atol, then return lhs

If x[feat_id] >= feat_value, then return rhs

Example

To represent the tree x[1] <= 0.0 ? -1 : (x[1] <= 1.0 ? 0 : 1), do:

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:1]);

julia> f = MathOptAI.BinaryDecisionTree{Float64,Int}(
           1,
           0.0,
           -1,
           MathOptAI.BinaryDecisionTree{Float64,Int}(1, 1.0, 0, 1),
       )
BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
1-element Vector{VariableRef}:
 moai_BinaryDecisionTree_value[1]

julia> formulation
BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]
├ variables [4]
│ ├ moai_BinaryDecisionTree_value[1]
│ ├ moai_BinaryDecisionTree_z[1]
│ ├ moai_BinaryDecisionTree_z[2]
│ └ moai_BinaryDecisionTree_z[3]
└ constraints [7]
  ├ moai_BinaryDecisionTree_z[1] + moai_BinaryDecisionTree_z[2] + moai_BinaryDecisionTree_z[3] = 1
  ├ moai_BinaryDecisionTree_z[1] --> {x[1] ≤ -1.0e-6}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≥ 0}
  ├ moai_BinaryDecisionTree_z[2] --> {x[1] ≤ 0.999999}
  ├ moai_BinaryDecisionTree_z[3] --> {x[1] ≥ 0}
  ├ moai_BinaryDecisionTree_z[3] --> {x[1] ≥ 1}
  └ moai_BinaryDecisionTree_z[1] - moai_BinaryDecisionTree_z[3] + moai_BinaryDecisionTree_value[1] = 0

Conv2d

MathOptAI.Conv2d — Type

Conv2d(
    weight::Array{T,4},
    bias::Vector{T};
    input_size::Tuple{Int,Int,Int},
    stride::Tuple{Int,Int} = (1, 1),
    padding::Tuple{Int,Int} = (0, 0),
) where {T} <: AbstractPredictor

An AbstractPredictor that represents a mean pooling layer a 2-dimensional convolutional layer.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[h in 1:2, w in 1:3])
2×3 Matrix{VariableRef}:
 x[1,1]  x[1,2]  x[1,3]
 x[2,1]  x[2,2]  x[2,3]

julia> weight = reshape(collect(1.0:8.0), 2, 2, 1, 2);

julia> bias = [-1.0, -2.0];

julia> predictor = MathOptAI.Conv2d(weight, bias; input_size = (2, 3, 1))
Conv2d{Float64}((2, 3, 1), [1.0 3.0; 2.0 4.0;;;; 5.0 7.0; 6.0 8.0], [-1.0, -2.0], (1, 1), (0, 0))

julia> y, formulation = MathOptAI.add_predictor(model, predictor, vec(x));

julia> y
4-element Vector{VariableRef}:
 moai_Conv2d[1]
 moai_Conv2d[2]
 moai_Conv2d[3]
 moai_Conv2d[4]

julia> formulation
Conv2d{Float64}((2, 3, 1), [1.0 3.0; 2.0 4.0;;;; 5.0 7.0; 6.0 8.0], [-1.0, -2.0], (1, 1), (0, 0))
├ variables [4]
│ ├ moai_Conv2d[1]
│ ├ moai_Conv2d[2]
│ ├ moai_Conv2d[3]
│ └ moai_Conv2d[4]
└ constraints [4]
  ├ -4 x[1,1] - 3 x[2,1] - 2 x[1,2] - x[2,2] + moai_Conv2d[1] = -1
  ├ -4 x[1,2] - 3 x[2,2] - 2 x[1,3] - x[2,3] + moai_Conv2d[2] = -1
  ├ -8 x[1,1] - 7 x[2,1] - 6 x[1,2] - 5 x[2,2] + moai_Conv2d[3] = -2
  └ -8 x[1,2] - 7 x[2,2] - 6 x[1,3] - 5 x[2,3] + moai_Conv2d[4] = -2

julia> y, formulation =
           MathOptAI.add_predictor(model, predictor, vec(x); reduced_space = true);

julia> y
4-element Vector{AffExpr}:
 4 x[1,1] + 2 x[1,2] + 3 x[2,1] + x[2,2] - 1
 4 x[1,2] + 2 x[1,3] + 3 x[2,2] + x[2,3] - 1
 8 x[1,1] + 6 x[1,2] + 7 x[2,1] + 5 x[2,2] - 2
 8 x[1,2] + 6 x[1,3] + 7 x[2,2] + 5 x[2,3] - 2

julia> formulation
ReducedSpace(Conv2d{Float64}((2, 3, 1), [1.0 3.0; 2.0 4.0;;;; 5.0 7.0; 6.0 8.0], [-1.0, -2.0], (1, 1), (0, 0)))
├ variables [0]
└ constraints [0]

GCNConv

MathOptAI.GCNConv — Type

GCNConv(;
    weights::Matrix{T},
    bias::Vector{T},
    edge_index::Vector{Pair{Int,Int}},
)

An AbstractPredictor that represents a graph convolutional network operator:

\[Y = D^{-1/2} A D^{-1/2} X W + b\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:3, 1:2]);

julia> f = MathOptAI.GCNConv(;
           weights = [1.0; 2.0;;],
           bias = [7.0],
           edge_index = [1 => 2, 2 => 1, 2 => 3, 3 => 2],
       );

julia> y, formulation = MathOptAI.add_predictor(model, f, vec(x));

julia> y
3-element Vector{VariableRef}:
 moai_GCNConv[1]
 moai_GCNConv[2]
 moai_GCNConv[3]

julia> formulation
GCNConv{Float64}([1.0; 2.0;;], [7.0], [0.4999999999999999 0.40824829046386296 0.0; 0.40824829046386296 0.33333333333333337 0.40824829046386296; 0.0 0.40824829046386296 0.4999999999999999])
├ variables [3]
│ ├ moai_GCNConv[1]
│ ├ moai_GCNConv[2]
│ └ moai_GCNConv[3]
└ constraints [3]
  ├ -0.4999999999999999 x[1,1] - 0.40824829046386296 x[2,1] - x[1,2] - 0.8164965809277259 x[2,2] + moai_GCNConv[1] = 7
  ├ -0.40824829046386296 x[1,1] - 0.33333333333333337 x[2,1] - 0.40824829046386296 x[3,1] - 0.8164965809277259 x[1,2] - 0.6666666666666667 x[2,2] - 0.8164965809277259 x[3,2] + moai_GCNConv[2] = 7
  └ -0.40824829046386296 x[2,1] - 0.4999999999999999 x[3,1] - 0.8164965809277259 x[2,2] - x[3,2] + moai_GCNConv[3] = 7

GELU

MathOptAI.GELU — Type

GeLU() <: AbstractPredictor

An AbstractPredictor representing the Gaussian Error Linear Units function:

\[y \approx x * (1 + \tanh(\sqrt(2 / \pi) * (x + 0.044715 x^3))) / 2\]

as a smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.GELU()
GELU()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_GELU[1]
 moai_GELU[2]

julia> formulation
GELU()
├ variables [2]
│ ├ moai_GELU[1]
│ └ moai_GELU[2]
└ constraints [6]
  ├ moai_GELU[1] ≥ -0.17
  ├ moai_GELU[1] ≤ 0.8411919906082768
  ├ moai_GELU[1] - ((0.5 x[1]) * (1 + tanh(0.7978845608028654 * (x[1] + (0.044715 * (x[1] ^ 3)))))) = 0
  ├ moai_GELU[2] ≥ -0.17
  ├ moai_GELU[2] ≤ 1.954597694087775
  └ moai_GELU[2] - ((0.5 x[2]) * (1 + tanh(0.7978845608028654 * (x[2] + (0.044715 * (x[2] ^ 3)))))) = 0

GrayBox

MathOptAI.GrayBox — Type

 GrayBox(
    predictor::P;
    device::String = "cpu",
    hessian::Bool = true,
) where {P}

An AbstractPredictor that represents the relationship:

\[y = f(x)\]

as a vector nonlinear operator.

This predictor should not be used directly; it is intended to be used by extensions like Flux and PyTorch.

Example

julia> using JuMP, MathOptAI, Flux

julia> chain = Flux.Chain(Flux.Dense(1 => 16, Flux.relu), Flux.Dense(16 => 1));

julia> model = Model();

julia> @variable(model, x[1:1]);

julia> y, _ = MathOptAI.add_predictor(model, chain, x; gray_box = true);

julia> y
1-element Vector{VariableRef}:
 moai_Flux[1]

julia> print(model)
Feasibility
Subject to
 [x[1], moai_Flux[1]] ∈ VectorNonlinearOracle{Float64}(;
     dimension = 2,
     l = [0.0],
     u = [0.0],
     ...,
 )

LayerNorm

MathOptAI.LayerNorm — Type

LayerNorm(
    shape::NTuple{N,Int};
    input_size::Tuple{Int,Int,Int},
    eps::Float64 = 1e-5,
) where {N}

An AbstractPredictor that represents the relationship:

\[y = \frac{x - E[x]}{\sqrt{Var(x) + \eps}}\]

where E and Var are computed over the first shape dimensions of x.

!!! Note This layer does not implement the affine scaling seen in some layers. Apply Scale to the outputs of this predictor instead.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2, 1:3]);

julia> f = MathOptAI.LayerNorm(
           (2,);
           eps = 0.0,
           input_size = (2, 3, 1),
           weight = [1.0, 2.0],
           bias = [0.5, 0.6],
       )
LayerNorm{Float64, 1}((2, 3, 1), (2,), 0.0, [1.0, 2.0], [0.5, 0.6])

julia> y, formulation = MathOptAI.add_predictor(model, f, vec(x));

julia> y
6-element Vector{VariableRef}:
 moai_LayerNorm[1]
 moai_LayerNorm[2]
 moai_LayerNorm[3]
 moai_LayerNorm[4]
 moai_LayerNorm[5]
 moai_LayerNorm[6]

julia> formulation
LayerNorm{Float64, 1}((2, 3, 1), (2,), 0.0, [1.0, 2.0], [0.5, 0.6])
├ variables [3]
│ ├ VariableRef[moai_LayerNorm_μ[1], moai_LayerNorm_μ[2], moai_LayerNorm_μ[3]]
│ ├ VariableRef[moai_LayerNorm_σ[1], moai_LayerNorm_σ[2], moai_LayerNorm_σ[3]]
│ └ VariableRef[moai_LayerNorm[1], moai_LayerNorm[2], moai_LayerNorm[3], moai_LayerNorm[4], moai_LayerNorm[5], moai_LayerNorm[6]]
└ constraints [15]
  ├ moai_LayerNorm_σ[1] ≥ 0
  ├ moai_LayerNorm_σ[2] ≥ 0
  ├ moai_LayerNorm_σ[3] ≥ 0
  ├ -x[1,1] - x[2,1] + 2 moai_LayerNorm_μ[1] = 0
  ├ -0.5 x[1,1]² + moai_LayerNorm_μ[1]*x[1,1] - 0.5 x[2,1]² + moai_LayerNorm_μ[1]*x[2,1] - moai_LayerNorm_μ[1]² + moai_LayerNorm_σ[1]² = 0
  ├ moai_LayerNorm_σ[1]*moai_LayerNorm[1] - x[1,1] + moai_LayerNorm_μ[1] - 0.5 moai_LayerNorm_σ[1] = 0
  ├ moai_LayerNorm_σ[1]*moai_LayerNorm[2] - 2 x[2,1] + 2 moai_LayerNorm_μ[1] - 0.6 moai_LayerNorm_σ[1] = 0
  ├ -x[1,2] - x[2,2] + 2 moai_LayerNorm_μ[2] = 0
  ├ -0.5 x[1,2]² + moai_LayerNorm_μ[2]*x[1,2] - 0.5 x[2,2]² + moai_LayerNorm_μ[2]*x[2,2] - moai_LayerNorm_μ[2]² + moai_LayerNorm_σ[2]² = 0
  ├ moai_LayerNorm_σ[2]*moai_LayerNorm[3] - x[1,2] + moai_LayerNorm_μ[2] - 0.5 moai_LayerNorm_σ[2] = 0
  ├ moai_LayerNorm_σ[2]*moai_LayerNorm[4] - 2 x[2,2] + 2 moai_LayerNorm_μ[2] - 0.6 moai_LayerNorm_σ[2] = 0
  ├ -x[1,3] - x[2,3] + 2 moai_LayerNorm_μ[3] = 0
  ├ -0.5 x[1,3]² + moai_LayerNorm_μ[3]*x[1,3] - 0.5 x[2,3]² + moai_LayerNorm_μ[3]*x[2,3] - moai_LayerNorm_μ[3]² + moai_LayerNorm_σ[3]² = 0
  ├ moai_LayerNorm_σ[3]*moai_LayerNorm[5] - x[1,3] + moai_LayerNorm_μ[3] - 0.5 moai_LayerNorm_σ[3] = 0
  └ moai_LayerNorm_σ[3]*moai_LayerNorm[6] - 2 x[2,3] + 2 moai_LayerNorm_μ[3] - 0.6 moai_LayerNorm_σ[3] = 0

LeakyReLU

MathOptAI.LeakyReLU — Type

LeakyReLU(;
    negative_slope::Float64,
    relu::AbstractPredictor = ReLU(),
) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \max\{0, x\} + \eta \cdot \min\{0, x\}\]

or equivalently:

\[y = \eta \cdot x + (1 - \eta) \cdot \max\{0, x\}\]

where negative_slope is $\eta$ and relu is used to represent $\max\{0, x\}$.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.LeakyReLU(; negative_slope = 0.01)
LeakyReLU{ReLU}(0.01, ReLU())

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_LeakyReLU[1]
 moai_LeakyReLU[2]

julia> formulation
LeakyReLU{ReLU}(0.01, ReLU())
├ variables [4]
│ ├ moai_ReLU[1]
│ ├ moai_ReLU[2]
│ ├ moai_LeakyReLU[1]
│ └ moai_LeakyReLU[2]
└ constraints [8]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] ≤ 1
  ├ moai_ReLU[1] - max(0, x[1]) = 0
  ├ moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] ≤ 2
  ├ moai_ReLU[2] - max(0, x[2]) = 0
  ├ -0.01 x[1] + moai_LeakyReLU[1] - 0.99 moai_ReLU[1] = 0
  └ -0.01 x[2] + moai_LeakyReLU[2] - 0.99 moai_ReLU[2] = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 (0.01 x[1]) + (0.99 * max(0, x[1]))
 (0.01 x[2]) + (0.99 * max(0, x[2]))

julia> formulation
ReducedSpace(LeakyReLU{ReLU}(0.01, ReLU()))
├ variables [0]
└ constraints [0]

MaxPool2d

MathOptAI.MaxPool2d — Type

MaxPool2d(
    kernel_size::Tuple{Int,Int};
    input_size::Tuple{Int,Int,Int},
    stride::Tuple{Int,Int} = (1, 1),
    padding::Tuple{Int,Int} = (0, 0),
) <: AbstractPredictor

An AbstractPredictor that represents a two-dimensinal max pooling layer.

The max function is implemented as a non-smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[h in 1:2, w in 1:4])
2×4 Matrix{VariableRef}:
 x[1,1]  x[1,2]  x[1,3]  x[1,4]
 x[2,1]  x[2,2]  x[2,3]  x[2,4]

julia> predictor = MathOptAI.MaxPool2d((2, 2); input_size = (2, 4, 1))
MaxPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0))

julia> y, formulation = MathOptAI.add_predictor(model, predictor, vec(x));

julia> y
2-element Vector{VariableRef}:
 moai_MaxPool2d[1]
 moai_MaxPool2d[2]

julia> formulation
MaxPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0))
├ variables [2]
│ ├ moai_MaxPool2d[1]
│ └ moai_MaxPool2d[2]
└ constraints [2]
  ├ moai_MaxPool2d[1] - max(max(max(x[1,1], x[2,1]), x[1,2]), x[2,2]) = 0
  └ moai_MaxPool2d[2] - max(max(max(x[1,3], x[2,3]), x[1,4]), x[2,4]) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, predictor, vec(x); reduced_space = true);

julia> y
2-element Vector{NonlinearExpr}:
 max(max(max(x[1,1], x[2,1]), x[1,2]), x[2,2])
 max(max(max(x[1,3], x[2,3]), x[1,4]), x[2,4])

julia> formulation
ReducedSpace(MaxPool2d((2, 4, 1), (2, 2), (2, 2), (0, 0)))
├ variables [0]
└ constraints [0]

MaxPool2dBigM

MathOptAI.MaxPool2dBigM — Type

MaxPool2dBigM(
    kernel_size::Tuple{Int,Int};
    M::Float64,
    input_size::Tuple{Int,Int,Int},
    stride::Tuple{Int,Int} = kernel_size,
    padding::Tuple{Int,Int} = (0, 0),
) <: AbstractPredictor

An AbstractPredictor that represents a two-dimensinal max pooling layer.

The max function is implemented as a big-M mixed-integer linear program.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[h in 1:2, w in 1:4])
2×4 Matrix{VariableRef}:
 x[1,1]  x[1,2]  x[1,3]  x[1,4]
 x[2,1]  x[2,2]  x[2,3]  x[2,4]

julia> predictor =
           MathOptAI.MaxPool2dBigM((2, 2); input_size = (2, 4, 1), M = 100.0)
MaxPool2dBigM((2, 4, 1), (2, 2), (2, 2), (0, 0), 100.0)

julia> y, formulation = MathOptAI.add_predictor(model, predictor, vec(x));

julia> y
2-element Vector{VariableRef}:
 moai_MaxPool2d[1]
 moai_MaxPool2d[2]

julia> formulation
MaxPool2dBigM((2, 4, 1), (2, 2), (2, 2), (0, 0), 100.0)
├ variables [10]
│ ├ moai_MaxPool2d[1]
│ ├ moai_MaxPool2d[2]
│ ├ moai_z[h=1,w=1,c=1][1]
│ ├ moai_z[h=1,w=1,c=1][2]
│ ├ moai_z[h=1,w=1,c=1][3]
│ ├ moai_z[h=1,w=1,c=1][4]
│ ├ moai_z[h=1,w=2,c=1][1]
│ ├ moai_z[h=1,w=2,c=1][2]
│ ├ moai_z[h=1,w=2,c=1][3]
│ └ moai_z[h=1,w=2,c=1][4]
└ constraints [18]
  ├ -x[1,1] + moai_MaxPool2d[1] ≥ 0
  ├ -x[1,1] + moai_MaxPool2d[1] - 100 moai_z[h=1,w=1,c=1][1] ≤ 0
  ├ -x[1,2] + moai_MaxPool2d[1] ≥ 0
  ├ -x[1,2] + moai_MaxPool2d[1] - 100 moai_z[h=1,w=1,c=1][3] ≤ 0
  ├ -x[2,1] + moai_MaxPool2d[1] ≥ 0
  ├ -x[2,1] + moai_MaxPool2d[1] - 100 moai_z[h=1,w=1,c=1][2] ≤ 0
  ├ -x[2,2] + moai_MaxPool2d[1] ≥ 0
  ├ -x[2,2] + moai_MaxPool2d[1] - 100 moai_z[h=1,w=1,c=1][4] ≤ 0
  ├ moai_z[h=1,w=1,c=1][1] + moai_z[h=1,w=1,c=1][2] + moai_z[h=1,w=1,c=1][3] + moai_z[h=1,w=1,c=1][4] = 3
  ├ -x[1,3] + moai_MaxPool2d[2] ≥ 0
  ├ -x[1,3] + moai_MaxPool2d[2] - 100 moai_z[h=1,w=2,c=1][1] ≤ 0
  ├ -x[1,4] + moai_MaxPool2d[2] ≥ 0
  ├ -x[1,4] + moai_MaxPool2d[2] - 100 moai_z[h=1,w=2,c=1][3] ≤ 0
  ├ -x[2,3] + moai_MaxPool2d[2] ≥ 0
  ├ -x[2,3] + moai_MaxPool2d[2] - 100 moai_z[h=1,w=2,c=1][2] ≤ 0
  ├ -x[2,4] + moai_MaxPool2d[2] ≥ 0
  ├ -x[2,4] + moai_MaxPool2d[2] - 100 moai_z[h=1,w=2,c=1][4] ≤ 0
  └ moai_z[h=1,w=2,c=1][1] + moai_z[h=1,w=2,c=1][2] + moai_z[h=1,w=2,c=1][3] + moai_z[h=1,w=2,c=1][4] = 3

Permutation

MathOptAI.Permutation — Type

Permutation(p::Vector{Int}) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = x[p]\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> f = MathOptAI.Permutation([2, 1])
Permutation([2, 1])

julia> y, formulation =
           MathOptAI.add_predictor(model, f, x; reduced_space = true);

julia> y
2-element Vector{VariableRef}:
 x[2]
 x[1]

Pipeline

MathOptAI.Pipeline — Type

Pipeline(layers::Vector{AbstractPredictor}) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = (l_1 \circ \ldots \circ l_N)(x)\]

where $l_i$ are a list of other AbstractPredictors.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> f = MathOptAI.Pipeline(
           MathOptAI.Affine([1.0 2.0], [0.0]),
           MathOptAI.ReLUQuadratic(),
       )
Pipeline with layers:
 * Affine(A, b) [input: 2, output: 1]
 * ReLUQuadratic(nothing)

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
1-element Vector{VariableRef}:
 moai_ReLU[1]

julia> formulation
Affine(A, b) [input: 2, output: 1]
├ variables [1]
│ └ moai_Affine[1]
└ constraints [1]
  └ x[1] + 2 x[2] - moai_Affine[1] = 0
ReLUQuadratic(nothing)
├ variables [2]
│ ├ moai_ReLU[1]
│ └ moai_z[1]
└ constraints [4]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_z[1] ≥ 0
  ├ moai_Affine[1] - moai_ReLU[1] + moai_z[1] = 0
  └ moai_ReLU[1]*moai_z[1] = 0

PytorchModel

MathOptAI.PytorchModel — Type

PytorchModel(filename::String)

A wrapper struct for loading a PyTorch model.

The only supported file extension is .pt, where the .pt file has been created using torch.save(model, filename).

Warning

To use PytorchModel, your code must load the PythonCall package:

import PythonCall

See the Python integration section of the documentation for instructions on how to link Python to Julia.

Example

julia> using MathOptAI

julia> using PythonCall  #  This line is important!

julia> predictor = PytorchModel("model.pt");

Quantile

MathOptAI.Quantile — Type

Quantile{D}(distribution::D, quantiles::Vector{Float64}) where {D}

An AbstractPredictor that represents the quantiles of distribution.

Example

julia> using JuMP, Distributions, MathOptAI

julia> model = Model();

julia> @variable(model, 1 <= x <= 2);

julia> predictor = MathOptAI.Quantile([0.1, 0.9]) do x
           return Distributions.Normal(x, 3 - x)
       end
Quantile(_, [0.1, 0.9])

julia> y, formulation = MathOptAI.add_predictor(model, predictor, [x]);

julia> y
2-element Vector{VariableRef}:
 moai_quantile[1]
 moai_quantile[2]

julia> formulation
Quantile(_, [0.1, 0.9])
├ variables [2]
│ ├ moai_quantile[1]
│ └ moai_quantile[2]
└ constraints [2]
  ├ moai_quantile[1] - op_quantile_0.1(x) = 0
  └ moai_quantile[2] - op_quantile_0.9(x) = 0

ReducedSpace

MathOptAI.ReducedSpace — Type

ReducedSpace(predictor::AbstractPredictor)

A wrapper type for other predictors that implement a reduced-space formulation.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> predictor = MathOptAI.ReducedSpace(MathOptAI.ReLU());

julia> y, formulation = MathOptAI.add_predictor(model, predictor, x);

julia> y
2-element Vector{NonlinearExpr}:
 max(0, x[1])
 max(0, x[2])

ReLU

MathOptAI.ReLU — Type

ReLU() <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \max\{0, x\}\]

as a non-smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.ReLU()
ReLU()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_ReLU[1]
 moai_ReLU[2]

julia> formulation
ReLU()
├ variables [2]
│ ├ moai_ReLU[1]
│ └ moai_ReLU[2]
└ constraints [6]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] ≤ 1
  ├ moai_ReLU[1] - max(0, x[1]) = 0
  ├ moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] ≤ 2
  └ moai_ReLU[2] - max(0, x[2]) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 max(0, x[1])
 max(0, x[2])

julia> formulation
ReducedSpace(ReLU())
├ variables [0]
└ constraints [0]

ReLUBigM

MathOptAI.ReLUBigM — Type

ReLUBigM(M::Float64) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \max\{0, x\}\]

via the big-M MIP reformulation:

\[\begin{aligned} y \ge 0 \\ y \ge x \\ y \le M z \\ y \le x + M(1 - z) \\ z \in\{0, 1\} \end{aligned}\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -3 <= x[i in 1:2] <= i);

julia> f = MathOptAI.ReLUBigM(100.0)
ReLUBigM(100.0)

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_ReLU[1]
 moai_ReLU[2]

julia> formulation
ReLUBigM(100.0)
├ variables [4]
│ ├ moai_ReLU[1]
│ ├ moai_ReLU[2]
│ ├ moai_z[1]
│ └ moai_z[2]
└ constraints [12]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] ≤ 1
  ├ moai_z[1] binary
  ├ -x[1] + moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] - moai_z[1] ≤ 0
  ├ -x[1] + moai_ReLU[1] + 3 moai_z[1] ≤ 3
  ├ moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] ≤ 2
  ├ moai_z[2] binary
  ├ -x[2] + moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] - 2 moai_z[2] ≤ 0
  └ -x[2] + moai_ReLU[2] + 3 moai_z[2] ≤ 3

ReLUQuadratic

MathOptAI.ReLUQuadratic — Type

ReLUQuadratic(; relaxation_parameter = nothing) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \max\{0, x\}\]

by the reformulation:

\[\begin{aligned} x = y - z \\ y \cdot z = 0 \\ y, z \ge 0 \end{aligned}\]

If relaxation_parameter is set to a value ϵ, the constraints become:

\[\begin{aligned} x = y - z \\ y \cdot z \leq \epsilon \\ y, z \ge 0 \end{aligned}\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.ReLUQuadratic()
ReLUQuadratic(nothing)

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_ReLU[1]
 moai_ReLU[2]

julia> formulation
ReLUQuadratic(nothing)
├ variables [4]
│ ├ moai_ReLU[1]
│ ├ moai_ReLU[2]
│ ├ moai_z[1]
│ └ moai_z[2]
└ constraints [12]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] ≤ 1
  ├ moai_z[1] ≥ 0
  ├ moai_z[1] ≤ 1
  ├ x[1] - moai_ReLU[1] + moai_z[1] = 0
  ├ moai_ReLU[1]*moai_z[1] = 0
  ├ moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] ≤ 2
  ├ moai_z[2] ≥ 0
  ├ moai_z[2] ≤ 1
  ├ x[2] - moai_ReLU[2] + moai_z[2] = 0
  └ moai_ReLU[2]*moai_z[2] = 0

ReLUSOS1

MathOptAI.ReLUSOS1 — Type

ReLUSOS1() <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \max\{0, x\}\]

by the reformulation:

\[\begin{aligned} x = y - z \\ [y, z] \in SOS1 \\ y, z \ge 0 \end{aligned}\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.ReLUSOS1()
ReLUSOS1()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_ReLU[1]
 moai_ReLU[2]

julia> formulation
ReLUSOS1()
├ variables [4]
│ ├ moai_ReLU[1]
│ ├ moai_ReLU[2]
│ ├ moai_z[1]
│ └ moai_z[2]
└ constraints [12]
  ├ moai_ReLU[1] ≥ 0
  ├ moai_ReLU[1] ≤ 1
  ├ moai_z[1] ≥ 0
  ├ moai_z[1] ≤ 1
  ├ x[1] - moai_ReLU[1] + moai_z[1] = 0
  ├ [moai_ReLU[1], moai_z[1]] ∈ MathOptInterface.SOS1{Float64}([1.0, 2.0])
  ├ moai_ReLU[2] ≥ 0
  ├ moai_ReLU[2] ≤ 2
  ├ moai_z[2] ≥ 0
  ├ moai_z[2] ≤ 1
  ├ x[2] - moai_ReLU[2] + moai_z[2] = 0
  └ [moai_ReLU[2], moai_z[2]] ∈ MathOptInterface.SOS1{Float64}([1.0, 2.0])

Scale

MathOptAI.Scale — Type

Scale(
    scale::Vector{T},
    bias::Vector{T},
) where {T} <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = Diag(scale)x + bias\]

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, 0 <= x[i in 1:2] <= i);

julia> f = MathOptAI.Scale([2.0, 3.0], [4.0, 5.0])
Scale(scale, bias)

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_Scale[1]
 moai_Scale[2]

julia> formulation
Scale(scale, bias)
├ variables [2]
│ ├ moai_Scale[1]
│ └ moai_Scale[2]
└ constraints [6]
  ├ moai_Scale[1] ≥ 4
  ├ moai_Scale[1] ≤ 6
  ├ moai_Scale[2] ≥ 5
  ├ moai_Scale[2] ≤ 11
  ├ 2 x[1] - moai_Scale[1] = -4
  └ 3 x[2] - moai_Scale[2] = -5

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{AffExpr}:
 2 x[1] + 4
 3 x[2] + 5

julia> formulation
ReducedSpace(Scale(scale, bias))
├ variables [0]
└ constraints [0]

Sigmoid

MathOptAI.Sigmoid — Type

Sigmoid() <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \frac{1}{1 + e^{-x}}\]

as a smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.Sigmoid()
Sigmoid()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_Sigmoid[1]
 moai_Sigmoid[2]

julia> formulation
Sigmoid()
├ variables [2]
│ ├ moai_Sigmoid[1]
│ └ moai_Sigmoid[2]
└ constraints [6]
  ├ moai_Sigmoid[1] ≥ 0.2689414213699951
  ├ moai_Sigmoid[1] ≤ 0.7310585786300049
  ├ moai_Sigmoid[1] - (1 / (1 + exp(-x[1]))) = 0
  ├ moai_Sigmoid[2] ≥ 0.2689414213699951
  ├ moai_Sigmoid[2] ≤ 0.8807970779778823
  └ moai_Sigmoid[2] - (1 / (1 + exp(-x[2]))) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 1 / (1 + exp(-x[1]))
 1 / (1 + exp(-x[2]))

julia> formulation
ReducedSpace(Sigmoid())
├ variables [0]
└ constraints [0]

SoftMax

MathOptAI.SoftMax — Type

SoftMax() <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \frac{e^{x}}{||e^{x}||_1}\]

as a smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> f = MathOptAI.SoftMax()
SoftMax()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_SoftMax[1]
 moai_SoftMax[2]

julia> formulation
SoftMax()
├ variables [3]
│ ├ moai_SoftMax_denom[1]
│ ├ moai_SoftMax[1]
│ └ moai_SoftMax[2]
└ constraints [8]
  ├ moai_SoftMax_denom[1] ≥ 0
  ├ moai_SoftMax_denom[1] - (0 + exp(x[2]) + exp(x[1])) = 0
  ├ moai_SoftMax[1] ≥ 0
  ├ moai_SoftMax[1] ≤ 1
  ├ moai_SoftMax[1] - (exp(x[1]) / moai_SoftMax_denom[1]) = 0
  ├ moai_SoftMax[2] ≥ 0
  ├ moai_SoftMax[2] ≤ 1
  └ moai_SoftMax[2] - (exp(x[2]) / moai_SoftMax_denom[1]) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 exp(x[1]) / moai_SoftMax_denom[1]
 exp(x[2]) / moai_SoftMax_denom[1]

julia> formulation
ReducedSpace(SoftMax())
├ variables [1]
│ └ moai_SoftMax_denom[1]
└ constraints [2]
  ├ moai_SoftMax_denom[1] ≥ 0
  └ moai_SoftMax_denom[1] - (0 + exp(x[2]) + exp(x[1])) = 0

SoftPlus

MathOptAI.SoftPlus — Type

SoftPlus(; beta = 1.0) <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \frac{1}{\beta} \log(1 + e^{\beta x})\]

as a smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.SoftPlus(; beta = 2.0)
SoftPlus(2.0)

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_SoftPlus[1]
 moai_SoftPlus[2]

julia> formulation
SoftPlus(2.0)
├ variables [2]
│ ├ moai_SoftPlus[1]
│ └ moai_SoftPlus[2]
└ constraints [6]
  ├ moai_SoftPlus[1] ≥ 0.0634640055214863
  ├ moai_SoftPlus[1] ≤ 1.0634640055214863
  ├ moai_SoftPlus[1] - (log(1 + exp(2 x[1])) / 2) = 0
  ├ moai_SoftPlus[2] ≥ 0.0634640055214863
  ├ moai_SoftPlus[2] ≤ 2.0090749639589047
  └ moai_SoftPlus[2] - (log(1 + exp(2 x[2])) / 2) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 log(1 + exp(2 x[1])) / 2
 log(1 + exp(2 x[2])) / 2

julia> formulation
ReducedSpace(SoftPlus(2.0))
├ variables [0]
└ constraints [0]

TAGConv

MathOptAI.TAGConv — Type

TAGConv(;
    weights::Vector{Matrix{T}},
    bias::Vector{T},
    edge_index::Vector{Pair{Int,Int}},
)

An AbstractPredictor that represents a topology adaptive graph convolutional network operator:

\[Y = \sum\limits_{k=0}^K (D^{-1/2} A D^{-1/2})^k X W_k + b\]

where:

$W_k$ is weights[k+1]

$A$ is the adjacency matrix constructed from edge_index

$b$ is the bias vector bias

$D$ is the diagonal degree matrix.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, x[1:3, 1:2]);

julia> f = MathOptAI.TAGConv(;
           weights = Matrix{Float64}[[1 2]', [3 4]', [5 6]'],
           bias = [7.0],
           edge_index = [1 => 2, 2 => 1, 2 => 3, 3 => 2],
       );

julia> y, formulation = MathOptAI.add_predictor(model, f, vec(x));

julia> y
3-element Vector{VariableRef}:
 moai_TAGConv[1]
 moai_TAGConv[2]
 moai_TAGConv[3]

julia> formulation
TAGConv{Float64}([[1.0; 2.0;;], [3.0; 4.0;;], [5.0; 6.0;;]], [7.0], [0.0 0.7071067811865475 0.0; 0.7071067811865475 0.0 0.7071067811865475; 0.0 0.7071067811865475 0.0])
├ variables [3]
│ ├ moai_TAGConv[1]
│ ├ moai_TAGConv[2]
│ └ moai_TAGConv[3]
└ constraints [3]
  ├ -3.4999999999999996 x[1,1] - 2.1213203435596424 x[2,1] - 2.4999999999999996 x[3,1] - 4.999999999999999 x[1,2] - 2.82842712474619 x[2,2] - 2.999999999999999 x[3,2] + moai_TAGConv[1] = 7
  ├ -2.1213203435596424 x[1,1] - 5.999999999999999 x[2,1] - 2.1213203435596424 x[3,1] - 2.82842712474619 x[1,2] - 7.999999999999998 x[2,2] - 2.82842712474619 x[3,2] + moai_TAGConv[2] = 7
  └ -2.4999999999999996 x[1,1] - 2.1213203435596424 x[2,1] - 3.4999999999999996 x[3,1] - 2.999999999999999 x[1,2] - 2.82842712474619 x[2,2] - 4.999999999999999 x[3,2] + moai_TAGConv[3] = 7

Tanh

MathOptAI.Tanh — Type

Tanh() <: AbstractPredictor

An AbstractPredictor that represents the relationship:

\[y = \tanh(x)\]

as a smooth nonlinear constraint.

Example

julia> using JuMP, MathOptAI

julia> model = Model();

julia> @variable(model, -1 <= x[i in 1:2] <= i);

julia> f = MathOptAI.Tanh()
Tanh()

julia> y, formulation = MathOptAI.add_predictor(model, f, x);

julia> y
2-element Vector{VariableRef}:
 moai_Tanh[1]
 moai_Tanh[2]

julia> formulation
Tanh()
├ variables [2]
│ ├ moai_Tanh[1]
│ └ moai_Tanh[2]
└ constraints [6]
  ├ moai_Tanh[1] ≥ -0.7615941559557649
  ├ moai_Tanh[1] ≤ 0.7615941559557649
  ├ moai_Tanh[1] - tanh(x[1]) = 0
  ├ moai_Tanh[2] ≥ -0.7615941559557649
  ├ moai_Tanh[2] ≤ 0.9640275800758169
  └ moai_Tanh[2] - tanh(x[2]) = 0

julia> y, formulation =
           MathOptAI.add_predictor(model, MathOptAI.ReducedSpace(f), x);

julia> y
2-element Vector{NonlinearExpr}:
 tanh(x[1])
 tanh(x[2])

julia> formulation
ReducedSpace(Tanh())
├ variables [0]
└ constraints [0]

AbstractFormulation

MathOptAI.AbstractFormulation — Type

abstract type AbstractFormulation end

An abstract type representing different formulations.

Formulation

MathOptAI.Formulation — Type

struct Formulation{P<:AbstractPredictor} <: AbstractFormulation
    predictor::P
    variables::Vector{Any}
    constraints::Vector{Any}
end

Fields

predictor: the predictor object used to build the formulation

variables: a vector of new decision variables added to the model

constraints: a vector of new constraints added to the model

Check the docstring of the predictor for an explanation of the formulation and the order of the elements in .variables and .constraints.

PipelineFormulation

MathOptAI.PipelineFormulation — Type

struct PipelineFormulation{P<:AbstractPredictor} <: AbstractFormulation
    predictor::P
    layers::Vector{Any}
end

Fields

predictor: the predictor object used to build the formulation

layers: the formulation associated with each of the layers in the pipeline

AbstractGPs

MathOptAI.add_predictor — Method

MathOptAI.add_predictor(
    model::JuMP.AbstractModel,
    predictor::MathOptAI.Quantile{<:AbstractGPs.PosteriorGP},
    x::Vector,
)

Add the quantiles of a trained Gaussian Process from AbstractGPs.jl to model.

Example

julia> using JuMP, MathOptAI, AbstractGPs

julia> x_data = 2π .* (0.0:0.1:1.0);

julia> y_data = sin.(x_data);

julia> fx = AbstractGPs.GP(AbstractGPs.Matern32Kernel())(x_data, 0.1);

julia> p_fx = AbstractGPs.posterior(fx, y_data);

julia> model = Model();

julia> @variable(model, 1 <= x[1:1] <= 6, start = 3);

julia> predictor = MathOptAI.Quantile(p_fx, [0.1, 0.9]);

julia> y, _ = MathOptAI.add_predictor(model, predictor, x);

julia> y
2-element Vector{VariableRef}:
 moai_quantile[1]
 moai_quantile[2]

julia> @objective(model, Max, y[2] - y[1])
moai_quantile[2] - moai_quantile[1]

DecisionTree

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(
    predictor::Union{
        DecisionTree.Ensemble,
        DecisionTree.DecisionTreeClassifier,
        DecisionTree.Leaf,
        DecisionTree.Node,
        DecisionTree.Root,
    },
)

Convert a binary decision tree from DecisionTree.jl to a BinaryDecisionTree.

Example

julia> using JuMP, MathOptAI, DecisionTree

julia> truth(x::Vector) = x[1] <= 0.5 ? -2 : (x[2] <= 0.3 ? 3 : 4)
truth (generic function with 1 method)

julia> features = abs.(sin.((1:10) .* (3:4)'));

julia> size(features)
(10, 2)

julia> labels = truth.(Vector.(eachrow(features)));

julia> tree = DecisionTree.build_tree(labels, features)
Decision Tree
Leaves: 3
Depth:  2

julia> model = Model();

julia> @variable(model, 0 <= x[1:2] <= 1);

julia> y, _ = MathOptAI.add_predictor(model, tree, x);

julia> y
1-element Vector{VariableRef}:
 moai_BinaryDecisionTree_value[1]

julia> MathOptAI.build_predictor(tree)
BinaryDecisionTree{Float64,Int64} [leaves=3, depth=2]

EvoTrees

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(predictor::EvoTrees.EvoTree{L,1}) where {L}

Convert a boosted tree from EvoTrees.jl to an AffineCombination of BinaryDecisionTree.

Example

julia> using JuMP, MathOptAI, EvoTrees

julia> truth(x::Vector) = x[1] <= 0.5 ? -2 : (x[2] <= 0.3 ? 3 : 4)
truth (generic function with 1 method)

julia> x_train = abs.(sin.((1:10) .* (3:4)'));

julia> size(x_train)
(10, 2)

julia> y_train = truth.(Vector.(eachrow(x_train)));

julia> config = EvoTrees.EvoTreeRegressor(; nrounds = 3);

julia> tree = EvoTrees.fit(config; x_train, y_train);

julia> model = Model();

julia> @variable(model, 0 <= x[1:2] <= 1);

julia> y, _ = MathOptAI.add_predictor(model, tree, x);

julia> y
1-element Vector{VariableRef}:
 moai_AffineCombination[1]

julia> MathOptAI.build_predictor(tree)
AffineCombination
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
├ 1.0 * BinaryDecisionTree{Float64,Float64} [leaves=3, depth=2]
└ 1.0 * [2.0]

Flux

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(
    predictor::Flux.Chain;
    config::Dict = Dict{Any,Any}(),
    gray_box::Bool = false,
    hessian::Bool = gray_box,
    input_size::Union{Nothing,NTuple{N,Int}} = nothing,
)

Convert a trained neural network from Flux.jl to a Pipeline.

Supported layers

Flux.Conv

Flux.Dense

Flux.flatten

Flux.LayerNorm

Flux.MaxPool

Flux.MeanPool

Flux.Scale

Flux.softmax

Supported activation functions

Flux.relu

Flux.sigmoid

Flux.softplus

Flux.tanh

Keyword arguments

config: see the Config section below.

gray_box: if true, the neural network is added using a GrayBox formulation.

hessian: if true, the gray_box formulations compute the Hessian of the output using Flux.hessian. The default for hessian is true if gray_box is used.

input_size: to disambiguate the input and output sizes of matrix inputs, chains containing Conv, LayerNorm, MaxPool, and MeanPool layers must specify an initial input size.

Config

The config dictionary controls how layers in Flux are mapped to AbstractPredictors.

Supported keys and and example key-value pairs are:

Flux.MaxPool => (k; kwargs...) -> MathOptAI.MaxPool2dBigM(k; M = 10.0, kwargs...)

Flux.relu => MathOptAI.ReLU

Flux.sigmoid => MathOptAI.Sigmoid

Flux.softmax => MathOptAI.SoftMax

Flux.softplus => MathOptAI.SoftPlus

Flux.tanh => MathOptAI.Tanh

Example

julia> using JuMP, MathOptAI, Flux

julia> chain = Flux.Chain(Flux.Dense(1 => 16, Flux.relu), Flux.Dense(16 => 1));

julia> model = Model();

julia> @variable(model, x[1:1]);

julia> y, _ = MathOptAI.add_predictor(
           model,
           chain,
           x;
           config = Dict(Flux.relu => MathOptAI.ReLU),
       );

julia> y
1-element Vector{VariableRef}:
 moai_Affine[1]

julia> MathOptAI.build_predictor(
           chain;
           config = Dict(Flux.relu => MathOptAI.ReLU),
       )
Pipeline with layers:
 * Affine(A, b) [input: 1, output: 16]
 * ReLU()
 * Affine(A, b) [input: 16, output: 1]

julia> MathOptAI.build_predictor(
           chain;
           config = Dict(Flux.relu => MathOptAI.ReLUQuadratic),
       )
Pipeline with layers:
 * Affine(A, b) [input: 1, output: 16]
 * ReLUQuadratic(nothing)
 * Affine(A, b) [input: 16, output: 1]

GLM

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(
    predictor::GLM.GeneralizedLinearModel{
        GLM.GlmResp{Vector{Float64},GLM.Bernoulli{Float64},GLM.LogitLink},
    };
    sigmoid::MathOptAI.AbstractPredictor = MathOptAI.Sigmoid(),
)

Convert a trained logistic model from GLM.jl to a Pipeline layer.

Keyword arguments

sigmoid: the predictor to use for the sigmoid layer.

Example

julia> using JuMP, MathOptAI, GLM

julia> X, Y = rand(10, 2), rand(Bool, 10);

julia> predictor = GLM.glm(X, Y, GLM.Bernoulli());

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> y, _ = MathOptAI.add_predictor(
           model,
           predictor,
           x;
           sigmoid = MathOptAI.Sigmoid(),
       );

julia> y
1-element Vector{VariableRef}:
 moai_Sigmoid[1]

julia> MathOptAI.build_predictor(predictor)
Pipeline with layers:
 * Affine(A, b) [input: 2, output: 1]
 * Sigmoid()

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(predictor::GLM.LinearModel)

Convert a trained linear model from GLM.jl to an Affine layer.

Example

julia> using JuMP, MathOptAI, GLM

julia> X, Y = rand(10, 2), rand(10);

julia> predictor = GLM.lm(X, Y);

julia> model = Model();

julia> @variable(model, x[1:2]);

julia> y, _ = MathOptAI.add_predictor(model, predictor, x);

julia> y
1-element Vector{VariableRef}:
 moai_Affine[1]

julia> MathOptAI.build_predictor(predictor)
Affine(A, b) [input: 2, output: 1]

Lux

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(
    predictor::Tuple{<:Lux.Chain,<:NamedTuple,<:NamedTuple};
    config::Dict = Dict{Any,Any}(),
)

Convert a trained neural network from Lux.jl to a Pipeline.

Supported layers

Lux.Dense

Lux.Scale

Supported activation functions

Lux.relu

Lux.sigmoid

Lux.softplus

Lux.softmax

Lux.tanh

Keyword arguments

config: see the Config section below.

Config

The config dictionary controls how layers in Flux are mapped to AbstractPredictors.

Supported keys and and example key-value pairs are:

Lux.relu => MathOptAI.ReLU

Lux.sigmoid => MathOptAI.Sigmoid

Lux.sigmoid_fast => MathOptAI.Sigmoid

Lux.softmax => MathOptAI.SoftMax

Lux.softplus => MathOptAI.SoftPlus

Lux.tanh => MathOptAI.Tanh

Lux.tanh_fast => MathOptAI.Tanh

Example

julia> using JuMP, MathOptAI, Lux, Random

julia> rng = Random.MersenneTwister();

julia> chain = Lux.Chain(Lux.Dense(1 => 16, Lux.relu), Lux.Dense(16 => 1))
Chain(
    layer_1 = Dense(1 => 16, relu),               # 32 parameters
    layer_2 = Dense(16 => 1),                     # 17 parameters
)         # Total: 49 parameters,
          #        plus 0 states.

julia> parameters, state = Lux.setup(rng, chain);

julia> model = Model();

julia> @variable(model, x[1:1]);

julia> y, _ = MathOptAI.add_predictor(
           model,
           (chain, parameters, state),
           x;
           config = Dict(Lux.relu => MathOptAI.ReLU),
       );

julia> y
1-element Vector{VariableRef}:
 moai_Affine[1]

julia> MathOptAI.build_predictor(
           (chain, parameters, state);
           config = Dict(Lux.relu => MathOptAI.ReLU),
       )
Pipeline with layers:
 * Affine(A, b) [input: 1, output: 16]
 * ReLU()
 * Affine(A, b) [input: 16, output: 1]

julia> MathOptAI.build_predictor(
           (chain, parameters, state);
           config = Dict(Lux.relu => MathOptAI.ReLUQuadratic),
       )
Pipeline with layers:
 * Affine(A, b) [input: 1, output: 16]
 * ReLUQuadratic(nothing)
 * Affine(A, b) [input: 16, output: 1]

PythonCall

MathOptAI.GCNConv — Method

MathOptAI.GCNConv(
    layer::PythonCall.Py;
    edge_index::Vector{Pair{Int,Int}},
)

Create a GCNConv layer from a torch_geometric.nn.GCNConv layer.

See the Graph neural networks tutorial for details.

MathOptAI.TAGConv — Method

MathOptAI.TAGConv(
    layer::PythonCall.Py;
    edge_index::Vector{Pair{Int,Int}},
)

Create a TAGConv layer from a torch_geometric.nn.TAGConv layer.

See the Graph neural networks tutorial for details.

MathOptAI.build_predictor — Method

MathOptAI.build_predictor(
    predictor::MathOptAI.PytorchModel;
    config::Dict = Dict{Any,Any}(),
    gray_box::Bool = false,
    hessian::Bool = gray_box,
    device::String = "cpu",
    input_size::Union{Nothing,NTuple{N,Int}} = nothing,
)

Convert a trained neural network from PyTorch via PythonCall.jl to a Pipeline.

Supported layers

nn.AvgPool2d

nn.Conv2d

nn.Dropout

nn.Flatten

nn.GELU

nn.LayerNorm

nn.LeakyReLU

nn.Linear

nn.MaxPool2d

nn.ReLU

nn.Sequential

nn.Sigmoid

nn.Softmax

nn.Softplus

nn.Tanh

Note that nn.Dropout layers are skipped because we assume that the PyTorch model is being evaluated, not trained.

Keyword arguments

config: see the Config section below.

gray_box: if true, the neural network is added using a GrayBox formulation.

hessian: if true, the gray_box formulation computes the Hessian of the output using torch.func.hessian. The default for hessian is true if gray_box is used.

device: device used to construct PyTorch tensors, for example, "cuda" to run on an Nvidia GPU.

input_size: to disambiguate the input and output sizes of matrix inputs, models containing AvgPool2d, Conv2d, and MaxPool2d layers must specify an initial input size.

Config

The config dictionary controls how layers in PyTorch are mapped to AbstractPredictors.

Supported keys and and example key-value pairs are:

:GELU => MathOptAI.GELU

:MaxPool2d => (k; kwargs...) -> MathOptAI.MaxPool2dBigM(k; M = 10.0, kwargs...)

:ReLU => MathOptAI.ReLU

:Sigmoid => MathOptAI.Sigmoid

:SoftMax => MathOptAI.SoftMax

:SoftPlus => (; beta) -> MathOptAI.SoftPlus(; beta)

:Tanh => MathOptAI.Tanh

Note that :LeakyReLU is not supported. Use :ReLU to control how the inner ReLU is modeled.

StatsModels

MathOptAI.add_predictor — Method

MathOptAI.add_predictor(
    model::JuMP.AbstractModel,
    predictor::StatsModels.TableRegressionModel,
    x::DataFrames.DataFrame;
    kwargs...,
)

Add a trained regression model from StatsModels.jl to model, using the DataFrame x as input.

In most cases, predictor should be a GLM.jl predictor supported by MathOptAI, but trained using @formula and a DataFrame instead of the raw matrix input.

In general, x may have some columns that are constant (Float64) and some columns that are JuMP decision variables.

Keyword arguments

All keyword arguments are passed to the corresponding add_predictor of the GLM extension.

Example

julia> using DataFrames, GLM, JuMP, MathOptAI

julia> train_df = DataFrames.DataFrame(x1 = rand(10), x2 = rand(10));

julia> train_df.y = 1.0 .* train_df.x1 + 2.0 .* train_df.x2 .+ rand(10);

julia> predictor = GLM.lm(GLM.@formula(y ~ x1 + x2), train_df);

julia> model = Model();

julia> test_df = DataFrames.DataFrame(
           x1 = rand(6),
           x2 = @variable(model, [1:6]),
       );

julia> test_df.y, _ = MathOptAI.add_predictor(model, predictor, test_df);

julia> test_df.y
6-element Vector{VariableRef}:
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]
 moai_Affine[1]

Extensions

MathOptAI.add_variables — Function

add_variables(
    model::JuMP.AbstractModel,
    x::Vector,
    n::Int,
    base_name::String,
)::Vector

Add a vector of n variables to model with the base name base_name.

Extensions

This function is a hook for JuMP extensions to interact with MathOptAI.

Implement this method for subtypes of model and x as needed.

The default method is:

function add_variables(
    model::JuMP.AbstractModel,
    x::Vector,
    n::Int,
    base_name::String,
)
    return JuMP.@variable(model, [1:n], base_name = base_name)
end

MathOptAI.get_variable_bounds — Function

get_variable_bounds(x::JuMP.AbstractVariableRef)

Return a tuple corresponding to the (lower, upper) variable bounds of x.

If there is no bound, the value returned is missing.

Extensions

This function is a hook for JuMP extensions to interact with MathOptAI.

Implement this method for subtypes of x as needed.

MathOptAI.set_variable_bounds — Function

set_variable_bounds(
    cons::Vector{Any},
    x::JuMP.AbstractVariableRef,
    l::Any,
    u::Any;
    optional::Bool,
)

Set the bounds on x to l and u, and push! their corresponding constraint references to cons.

If l or u are missing, do not set the bound.

If optional = true, you may choose to silently skip setting the bounds because they are not required for correctness.

The type of l and u depends on get_variable_bounds.

Extensions

This function is a hook for JuMP extensions to interact with MathOptAI.

Implement this method for subtypes of x as needed.

MathOptAI.get_variable_start — Function

get_variable_start(x::JuMP.AbstractVariableRef)

Get the primal starting value of x, or return missing if one is not set.

The return value of this function is propogated through the various AbstractPredictors, and the primal start of new output variables is set using set_variable_start.

Extensions

This function is a hook for JuMP extensions to interact with MathOptAI.

Implement this method for subtypes of x as needed.

MathOptAI.set_variable_start — Function

set_variable_start(x::JuMP.AbstractVariableRef, start::Any)

Set the primal starting value of x to start, or do nothing if start is missing.

The input value start of this function is computed by propogating the primal start of the input variables (obtained with get_variable_start) through the various AbstractPredictors.

Extensions

This function is a hook for JuMP extensions to interact with MathOptAI.

Implement this method for subtypes of x and start as needed.

replace_weights_with_variables

MathOptAI.replace_weights_with_variables — Function

replace_weights_with_variables(
    model::JuMP.AbstractModel,
    predictor::AbstractPredictor,
)

Convert predictor with trained weights into a predictor in which the weights are JuMP decision variables.

This function is useful when you wish to use constrained optimization to train small to moderate neural networks.

Warning

This function is experimental and it may change in any future release. If you use this feature, please open a GitHub issue and let us know your thoughts.

Example

julia> using JuMP, Flux, MathOptAI

julia> chain = Flux.Chain(Flux.Dense(2 => 3), Flux.softmax);

julia> model = Model();

julia> @variable(model, x[i in 1:2] == i);

julia> predictor = MathOptAI.build_predictor(chain)
Pipeline with layers:
 * Affine(A, b) [input: 2, output: 3]
 * SoftMax()

julia> predictor = MathOptAI.replace_weights_with_variables(model, predictor)
Pipeline with layers:
 * Affine(A, b) [input: 2, output: 3]
 * SoftMax()

julia> y, _ = MathOptAI.add_predictor(model, predictor, x);